Buy_Scenario3 logic is created.
buy_scenario3_highest_rated.py finds the highest rated item in the subsection and selects the item.

Note: There are some items that have not been rated yet, NoSuchElementException is used to pass the error.

buy_scenario3_lowest_rated.py works as the highest logic to find and select the lowest rated item in the subsection.